internal/trace: extract trace policy to OT-agnostic package
