+++ internal/trace/ot/ot.go
@@ -7,37 +7,12 @@ package ot
 import (
 	"context"
 	"net/http"
-	"strconv"
-	"strings"
 
 	"github.com/opentracing-contrib/go-stdlib/nethttp"
 	"github.com/opentracing/opentracing-go"
-	"go.uber.org/atomic"
+
+	"github.com/sourcegraph/sourcegraph/internal/trace/policy"
 )
-
-type TracePolicy string
-
-const (
-	// TraceNone turns off tracing.
-	TraceNone TracePolicy = "none"
-
-	// TraceSelective turns on tracing only for requests with the X-Sourcegraph-Should-Trace header
-	// set to a truthy value.
-	TraceSelective TracePolicy = "selective"
-
-	// TraceAll turns on tracing for all requests.
-	TraceAll TracePolicy = "all"
-)
-
-var trPolicy = atomic.NewString(string(TraceNone))
-
-func SetTracePolicy(newTracePolicy TracePolicy) {
-	trPolicy.Store(string(newTracePolicy))
-}
-
-func GetTracePolicy() TracePolicy {
-	return TracePolicy(trPolicy.Load())
-}
 
 // HTTPMiddleware wraps the handler with the following:
 //
@@ -54,73 +29,21 @@ func HTTPMiddleware(h http.Handler, opts ...nethttp.MWOption) http.Handler {
 func MiddlewareWithTracer(tr opentracing.Tracer, h http.Handler, opts ...nethttp.MWOption) http.Handler {
 	nethttpMiddleware := nethttp.Middleware(tr, h, append([]nethttp.MWOption{
 		nethttp.MWSpanFilter(func(r *http.Request) bool {
-			return ShouldTrace(r.Context())
+			return policy.ShouldTrace(r.Context())
 		}),
 	}, opts...)...)
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		var trace bool
-		switch GetTracePolicy() {
-		case TraceSelective:
-			trace = requestWantsTracing(r)
-		case TraceAll:
+		switch policy.GetTracePolicy() {
+		case policy.TraceSelective:
+			trace = policy.RequestWantsTracing(r)
+		case policy.TraceAll:
 			trace = true
 		default:
 			trace = false
 		}
-		nethttpMiddleware.ServeHTTP(w, r.WithContext(WithShouldTrace(r.Context(), trace)))
+		nethttpMiddleware.ServeHTTP(w, r.WithContext(policy.WithShouldTrace(r.Context(), trace)))
 	})
-}
-
-const traceHeader = "X-Sourcegraph-Should-Trace"
-const traceQuery = "trace"
-
-// requestWantsTrace returns true if a request is opting into tracing either
-// via our HTTP Header or our URL Query.
-func requestWantsTracing(r *http.Request) bool {
-	// Prefer header over query param.
-	if v := r.Header.Get(traceHeader); v != "" {
-		b, _ := strconv.ParseBool(v)
-		return b
-	}
-	// PERF: Avoid parsing RawQuery if "trace=" is not present
-	if strings.Contains(r.URL.RawQuery, "trace=") {
-		v := r.URL.Query().Get(traceQuery)
-		b, _ := strconv.ParseBool(v)
-		return b
-	}
-	return false
-}
-
-// Transport wraps an underlying HTTP RoundTripper, injecting the X-Sourcegraph-Should-Trace header
-// into outgoing requests whenever the shouldTraceKey context value is true.
-type Transport struct {
-	http.RoundTripper
-}
-
-func (r *Transport) RoundTrip(req *http.Request) (*http.Response, error) {
-	req.Header.Set(traceHeader, strconv.FormatBool(ShouldTrace(req.Context())))
-	t := nethttp.Transport{RoundTripper: r.RoundTripper}
-	return t.RoundTrip(req)
-}
-
-type key int
-
-const (
-	shouldTraceKey key = iota
-)
-
-// ShouldTrace returns true if the shouldTraceKey context value is true.
-func ShouldTrace(ctx context.Context) bool {
-	v, ok := ctx.Value(shouldTraceKey).(bool)
-	if !ok {
-		return false
-	}
-	return v
-}
-
-// WithShouldTrace sets the shouldTraceKey context value.
-func WithShouldTrace(ctx context.Context, shouldTrace bool) context.Context {
-	return context.WithValue(ctx, shouldTraceKey, shouldTrace)
 }
 
 // GetTracer returns the tracer to use for the given context. If ShouldTrace returns true, it
@@ -133,7 +56,7 @@ func GetTracer(ctx context.Context) opentracing.Tracer {
 // it returns the NoopTracer. If it returns true and the passed-in tracer is not nil, it returns the
 // passed-in tracer. Otherwise, it returns the global tracer.
 func getTracer(ctx context.Context, tracer opentracing.Tracer) opentracing.Tracer {
-	if !ShouldTrace(ctx) {
+	if !policy.ShouldTrace(ctx) {
 		return opentracing.NoopTracer{}
 	}
 	if tracer == nil {
